Add files via upload
After pre-processing the data and handling outliers, I performed basic EDA on all variables to get a better understanding of the dataset and the problem at hand. Once I had an understanding of the background of the customers, I could move to the next step of analysis - looking at the customer journey and seeing where there were major bottlenecks for different customer segments.